requirements.txt
boto3==1.28.38
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
PyPDF2==3.0.1
langchain==0.0.267
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
transformers>=4.38.0
torch>=2.0.0
accelerators>=0.1.0
sentencepiece>=0.1.99
protobuf>=3.20.0
safetensors>=0.4.0
